Feature/0.9.4

+++ penyamanager/user/tablereservationview.cpp
@@ -7,7 +7,7 @@
 
 #include <commons/utils.h>
 #include <commons/singletons.h>
-#include "numitemdialog.h"
+#include <commons/numitemdialog.h>
 #include "tablereservationview.h"
 #include "ui_tablereservationview.h"
 
@@ -155,15 +155,23 @@ namespace PenyaManager {
             } else {
                 // this table is reserved
                 ReservationPtr pReservationPtr = tableReservationMapItem->second;
-                QString guestName = QString("[%1] %2 %3").arg(pReservationPtr->m_idMember).arg(pReservationPtr->m_memberName).arg(pReservationPtr->m_memberSurname);
-                this->ui->tableReservationTableWidget->setItem(rowCount, 3, new QTableWidgetItem(guestName));
                 this->ui->tableReservationTableWidget->setItem(rowCount, 4, new QTableWidgetItem(QString::number(pReservationPtr->m_guestNum)));
-                if (pReservationPtr->m_idMember == pMemberPtr->m_id) {
-                    // show cancel button action
-                    QPushButton *pCancelButton = new QPushButton(tr("Cancel"), this->ui->tableReservationTableWidget);
-                    this->connect(pCancelButton, &QPushButton::clicked, std::bind(&TableReservationView::on_cancelButton_clicked, this, pReservationPtr->m_reservationId, pReservationItemPtr->m_itemType));
-                    this->ui->tableReservationTableWidget->setCellWidget(rowCount, 5, pCancelButton);
+                // do not show cancel when reservation is from Admin
+                if (pReservationPtr->m_isAdmin)
+                {
+                    this->ui->tableReservationTableWidget->setItem(rowCount, 3, new QTableWidgetItem(QString(tr("BLOCKED"))));
+                } else
+                {
+                    QString guestName = QString("%1 %2").arg(pReservationPtr->m_memberName).arg(pReservationPtr->m_memberSurname);
+                    this->ui->tableReservationTableWidget->setItem(rowCount, 3, new QTableWidgetItem(guestName));
+                    if (pReservationPtr->m_idMember == pMemberPtr->m_id) {
+                        // show cancel button action
+                        QPushButton *pCancelButton = new QPushButton(tr("Cancel"), this->ui->tableReservationTableWidget);
+                        this->connect(pCancelButton, &QPushButton::clicked, std::bind(&TableReservationView::on_cancelButton_clicked, this, pReservationPtr->m_reservationId, pReservationItemPtr->m_itemType));
+                        this->ui->tableReservationTableWidget->setCellWidget(rowCount, 5, pCancelButton);
+                    }
                 }
+
             }
             rowCount++;
         }
@@ -208,7 +216,8 @@ namespace PenyaManager {
         {
             ReservationPtr pReservationPtr = *iter;
             tableReservationMap[pReservationPtr->m_idItem] = pReservationPtr;
-            if (pReservationPtr->m_idMember == pMemberPtr->m_id)
+            // has reservation when reservation is not Admin
+            if (pReservationPtr->m_idMember == pMemberPtr->m_id && !pReservationPtr->m_isAdmin)
             {
                 hasReservation = true;
             }
@@ -240,20 +249,21 @@ namespace PenyaManager {
             }
         }
 
+        bool isAdmin = false;
         QString title;
         switch (itemType)
         {
             case ReservationItemType::LunchTableType:
                 title = "Table reservation";
-                Singletons::m_pDAO->makeTableReservation(date, reservationType, guestNum, pCurrMemberPtr->m_id, itemId);
+                Singletons::m_pDAO->makeTableReservation(date, reservationType, guestNum, pCurrMemberPtr->m_id, itemId, isAdmin);
                 break;
             case ReservationItemType::OvenType:
                 title = "Oven reservation";
-                Singletons::m_pDAO->makeOvenReservation(date, reservationType, pCurrMemberPtr->m_id, itemId);
+                Singletons::m_pDAO->makeOvenReservation(date, reservationType, pCurrMemberPtr->m_id, itemId, isAdmin);
                 break;
             case ReservationItemType::FireplaceType:
                 title = "Fireplace reservation";
-                Singletons::m_pDAO->makeFireplaceReservation(date, reservationType, pCurrMemberPtr->m_id, itemId);
+                Singletons::m_pDAO->makeFireplaceReservation(date, reservationType, pCurrMemberPtr->m_id, itemId, isAdmin);
                 break;
             default:
                 break;
@@ -285,8 +295,10 @@ namespace PenyaManager {
                 break;
         }
         QMessageBox::information(this, title, "Reservation cancelled");
-        // Go to dashboard window
-        m_switchCentralWidgetCallback(WindowKey::kMemberDashboardWindowKey);
+        MemberPtr pCurrMemberPtr = Singletons::m_pCurrMember;
+        QDate date = this->ui->calendarWidget->selectedDate();
+        ReservationType reservationType = static_cast<ReservationType>(this->ui->reservationTypeButtonGroup->checkedId());
+        fillTableReservations(pCurrMemberPtr, date, reservationType);
     }
     //
     void TableReservationView::on_lunchButton_clicked(bool checked)
